requirements.txt
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "3.12"
anyio==4.2.0 ; python_version >= "3.9" and python_version < "3.12"
arxiv==2.1.0 ; python_version >= "3.9" and python_version < "3.12"
asyncio==3.4.3 ; python_version >= "3.9" and python_version < "3.12"
autogenstudio==0.0.44a0 ; python_version >= "3.9" and python_version < "3.12"
certifi==2024.2.2 ; python_version >= "3.9" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.9" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.12" and platform_system == "Windows"
diskcache==5.6.3 ; python_version >= "3.9" and python_version < "3.12"
distro==1.9.0 ; python_version >= "3.9" and python_version < "3.12"
docker==7.0.0 ; python_version >= "3.9" and python_version < "3.12"
exceptiongroup==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.109.2 ; python_version >= "3.9" and python_version < "3.12"
feedparser==6.0.10 ; python_version >= "3.9" and python_version < "3.12"
flaml==2.1.1 ; python_version >= "3.9" and python_version < "3.12"
greenlet==3.0.3 ; python_version >= "3.9" and python_version < "3.12"
h11==0.14.0 ; python_version >= "3.9" and python_version < "3.12"
httpcore==1.0.3 ; python_version >= "3.9" and python_version < "3.12"
httpx==0.26.0 ; python_version >= "3.9" and python_version < "3.12"
idna==3.6 ; python_version >= "3.9" and python_version < "3.12"
numpy==1.26.4 ; python_version >= "3.9" and python_version < "3.12"
openai==1.12.0 ; python_version >= "3.9" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.9" and python_version < "3.12"
pillow==10.2.0 ; python_version >= "3.9" and python_version < "3.12"
playwright==1.41.2 ; python_version >= "3.9" and python_version < "3.12"
pyautogen==0.2.14 ; python_version >= "3.9" and python_version < "3.12"
pyautogen[lmm]==0.2.14 ; python_version >= "3.9" and python_version < "3.12"
pydantic-core==2.16.2 ; python_version >= "3.9" and python_version < "3.12"
pydantic==2.6.1 ; python_version >= "3.9" and python_version < "3.12"
pyee==11.0.1 ; python_version >= "3.9" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.9" and python_version < "3.12"
pywin32==306 ; python_version >= "3.9" and python_version < "3.12" and sys_platform == "win32"
regex==2023.12.25 ; python_version >= "3.9" and python_version < "3.12"
replicate==0.23.1 ; python_version >= "3.9" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.9" and python_version < "3.12"
sgmllib3k==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "3.12"
starlette==0.36.3 ; python_version >= "3.9" and python_version < "3.12"
termcolor==2.4.0 ; python_version >= "3.9" and python_version < "3.12"
tiktoken==0.6.0 ; python_version >= "3.9" and python_version < "3.12"
tqdm==4.66.2 ; python_version >= "3.9" and python_version < "3.12"
typer==0.9.0 ; python_version >= "3.9" and python_version < "3.12"
typing-extensions==4.9.0 ; python_version >= "3.9" and python_version < "3.12"
urllib3==2.2.0 ; python_version >= "3.9" and python_version < "3.12"
uvicorn==0.27.1 ; python_version >= "3.9" and python_version < "3.12"
